refactor: consolidate array utility functions into Arrays module
