+++ src/utils/treePanel.ts
@@ -8,20 +8,6 @@ export type TreePanelView<T> = {
   reveal(element: T, options?: TreePanelRevealOptions): PromiseLike<void>;
 };
 
-export function sortItems<T>(items: readonly T[], compare: (left: T, right: T) => number): T[] {
-  return items.slice().sort(compare);
-}
-
-export function sortItemsInPlace<T>(items: T[], compare: (left: T, right: T) => number): boolean {
-  const sortedItems = sortItems(items, compare);
-  if (sortedItems.every((item, index) => item === items[index])) {
-    return false;
-  }
-
-  items.splice(0, items.length, ...sortedItems);
-  return true;
-}
-
 export async function focusTreeItem<T>(treeView: TreePanelView<T>, item: T): Promise<void> {
   await treeView.reveal(item, { focus: true, select: false });
 }
